done fron over view

+++ overallReports/posts/views.py
@@ -140,7 +140,10 @@ def post_update(request, slug=None):
 
 
 def post_delete(request, slug=None):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
     instance = get_object_or_404(Post, slug=slug)
+    # form = PostForm(request.POST or None, request.FILES or None, instance=instance)
     instance.delete()
     messages.success(request,"Successfully Deleted")
     return redirect("posts:list")
